ADD TO: pysat_func && Scratch
pysat_func: removed __init__ file as variables can be assigned at any
time, and do not need to be instantiated every time.
scratch: added scratch.py to showcase modularity of a particular test
case and it's functions. This will be used as a cheat sheet for PYSAT_UI
later on.