Giving more specific names to main running files
Given that func and ui are rather indescript names, I've elected to
change the way these are named. One file does all the backend stuff and
processes stuff from the front end. While the front end deals with
showing and displaying the UI correctly.

The reason for this is because of the way QThread works. Trying to fuse
these two files would work but it would lose the the context in which
these are working.

In other words it makes sense to know who controls the front and who
controls the back.